requirements.txt
python-docx
pyTelegramBotAPI
python-dotenv
SQLAlchemy
SQLAlchemy-file